python/requirements.txt
flask
flask_cors
spotipy
mysql-connector-python
dotenv